pyproject.toml
[project]
name = "indic-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keyboard>=0.13.5",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sarvamai>=0.1.15",
    "scipy>=1.16.1",
    "sounddevice>=0.5.2",
    "streamlit>=1.47.1",
]
